move eth_vault functions from utils to eth_vault
