+++ e2e/examples/eth_vault/evm.rs
@@ -1,21 +1,26 @@
 use std::error::Error;
 
-use alloy::primitives::{Address, U256};
+use std::str::FromStr;
+
+use alloy::{
+    hex::FromHex,
+    primitives::{Address, Bytes, U256},
+    sol_types::SolValue,
+};
 use log::info;
 use valence_chain_client_utils::{
     ethereum::EthereumClient,
     evm::{base_client::EvmBaseClient, request_provider_client::RequestProviderClient},
 };
+use valence_encoder_utils::libraries::cctp_transfer::solidity_types::CCTPTransferConfig;
 
-use valence_e2e::{
-    async_run,
-    utils::{
-        solidity_contracts::{
-            MockERC20,
-            ValenceVault::{self},
-        },
-        vault::setup_cctp_transfer,
+use crate::async_run;
+use valence_e2e::utils::{
+    solidity_contracts::{
+        CCTPTransfer, ERC1967Proxy, MockERC20, MockTokenMessenger,
+        ValenceVault::{self},
     },
+    vault::setup_vault_config,
 };
 
 pub fn mine_blocks(
@@ -95,7 +100,7 @@ pub fn log_eth_balances(
 #[derive(Clone, Debug)]
 pub struct EthereumProgramLibraries {
     pub cctp_forwarder: Address,
-    pub lite_processor: Address,
+    pub _lite_processor: Address,
     pub valence_vault: Address,
 }
 
@@ -138,8 +143,7 @@ pub fn setup_eth_libraries(
     eth_client: &EthereumClient,
     eth_admin_addr: Address,
     eth_strategist_addr: Address,
-    deposit_acc_addr: Address,
-    withdraw_acc_addr: Address,
+    eth_program_accounts: EthereumProgramAccounts,
     cctp_messenger_addr: Address,
     usdc_token_addr: Address,
     noble_inbound_ica_addr: String,
@@ -152,7 +156,7 @@ pub fn setup_eth_libraries(
         rt,
         eth_client,
         noble_inbound_ica_addr,
-        deposit_acc_addr,
+        eth_program_accounts.deposit,
         eth_admin_addr,
         eth_strategist_addr,
         usdc_token_addr,
@@ -170,22 +174,193 @@ pub fn setup_eth_libraries(
         )?;
 
     info!("Setting up Valence Vault...");
-    let vault_address = valence_e2e::utils::vault::setup_valence_vault(
+    let vault_address = setup_valence_vault(
         rt,
         eth_client,
         eth_strategist_addr,
         eth_accounts,
         eth_admin_addr,
-        deposit_acc_addr,
-        withdraw_acc_addr,
+        eth_program_accounts,
         usdc_token_addr,
     )?;
 
     let libraries = EthereumProgramLibraries {
         cctp_forwarder: cctp_forwarder_addr,
-        lite_processor: lite_processor_address,
+        _lite_processor: lite_processor_address,
         valence_vault: vault_address,
     };
 
     Ok(libraries)
+}
+
+/// sets up a Valence Vault on Ethereum with a proxy.
+/// approves deposit & withdraw accounts.
+pub fn setup_valence_vault(
+    rt: &tokio::runtime::Runtime,
+    eth_client: &EthereumClient,
+    eth_strategist_acc: Address,
+    eth_accounts: &[Address],
+    admin: Address,
+    eth_program_accounts: EthereumProgramAccounts,
+    vault_deposit_token_addr: Address,
+) -> Result<Address, Box<dyn Error>> {
+    let eth_rp = async_run!(rt, eth_client.get_request_provider().await.unwrap());
+
+    info!("deploying Valence Vault on Ethereum...");
+    let vault_config = setup_vault_config(
+        eth_accounts,
+        eth_program_accounts.deposit,
+        eth_program_accounts.withdraw,
+        eth_strategist_acc,
+    );
+
+    // First deploy the implementation contract
+    let implementation_tx = ValenceVault::deploy_builder(&eth_rp)
+        .into_transaction_request()
+        .from(admin);
+
+    let implementation_address = async_run!(
+        rt,
+        eth_client
+            .execute_tx(implementation_tx)
+            .await
+            .unwrap()
+            .contract_address
+            .unwrap()
+    );
+
+    info!("Vault deployed at: {implementation_address}");
+
+    let proxy_address = async_run!(rt, {
+        // Deploy the proxy contract
+        let proxy_tx = ERC1967Proxy::deploy_builder(&eth_rp, implementation_address, Bytes::new())
+            .into_transaction_request()
+            .from(admin);
+
+        let proxy_address = eth_client
+            .execute_tx(proxy_tx)
+            .await
+            .unwrap()
+            .contract_address
+            .unwrap();
+        info!("Proxy deployed at: {proxy_address}");
+        proxy_address
+    });
+
+    // Initialize the Vault
+    let vault = ValenceVault::new(proxy_address, &eth_rp);
+
+    async_run!(rt, {
+        let initialize_tx = vault
+            .initialize(
+                admin,                            // owner
+                vault_config.abi_encode().into(), // encoded config
+                vault_deposit_token_addr,         // underlying token
+                "Valence Test Vault".to_string(), // vault token name
+                "vTEST".to_string(),              // vault token symbol
+                U256::from(1e18), // placeholder, tbd what a reasonable value should be here
+            )
+            .into_transaction_request()
+            .from(admin);
+
+        eth_client.execute_tx(initialize_tx).await.unwrap();
+    });
+
+    info!("Approving vault for withdraw account...");
+    valence_e2e::utils::ethereum::valence_account::approve_library(
+        rt,
+        eth_client,
+        eth_program_accounts.withdraw,
+        proxy_address,
+    );
+
+    info!("Approving vault for deposit account...");
+    valence_e2e::utils::ethereum::valence_account::approve_library(
+        rt,
+        eth_client,
+        eth_program_accounts.deposit,
+        proxy_address,
+    );
+
+    Ok(proxy_address)
+}
+
+pub fn setup_mock_token_messenger(
+    rt: &tokio::runtime::Runtime,
+    eth_client: &EthereumClient,
+) -> Result<Address, Box<dyn Error>> {
+    let eth_rp = async_run!(rt, eth_client.get_request_provider().await.unwrap());
+
+    info!("deploying Mock Token Messenger lib on Ethereum...");
+
+    let messenger_tx = MockTokenMessenger::deploy_builder(eth_rp).into_transaction_request();
+
+    let messenger_rx = async_run!(rt, eth_client.execute_tx(messenger_tx).await.unwrap());
+
+    let messenger_address = messenger_rx.contract_address.unwrap();
+    info!("Mock CCTP Token Messenger deployed at: {messenger_address}");
+
+    Ok(messenger_address)
+}
+
+#[allow(clippy::too_many_arguments)]
+pub fn setup_cctp_transfer(
+    rt: &tokio::runtime::Runtime,
+    eth_client: &EthereumClient,
+    noble_recipient: String,
+    input_account: Address,
+    admin: Address,
+    processor: Address,
+    usdc_token_address: Address,
+    cctp_token_messenger_address: Address,
+) -> Result<Address, Box<dyn Error>> {
+    let eth_rp = async_run!(rt, eth_client.get_request_provider().await.unwrap());
+
+    info!("deploying CCTP Transfer lib on Ethereum...");
+
+    // Decode the bech32 address
+    let (_, data) = bech32::decode(&noble_recipient)?;
+    // Convert to hex
+    let address_hex = hex::encode(data);
+    // Pad with zeroes to 32 bytes
+    let padded_hex = format!("{:0>64}", address_hex);
+
+    let cctp_transer_cfg = CCTPTransferConfig {
+        amount: U256::ZERO,
+        mintRecipient: alloy_primitives_encoder::FixedBytes::<32>::from_hex(padded_hex)?,
+        inputAccount: alloy_primitives_encoder::Address::from_str(
+            input_account.to_string().as_str(),
+        )?,
+        destinationDomain: 4,
+        cctpTokenMessenger: alloy_primitives_encoder::Address::from_str(
+            cctp_token_messenger_address.to_string().as_str(),
+        )?,
+        transferToken: alloy_primitives_encoder::Address::from_str(
+            usdc_token_address.to_string().as_str(),
+        )?,
+    };
+
+    let cctp_tx = CCTPTransfer::deploy_builder(
+        &eth_rp,
+        admin,
+        processor,
+        alloy_sol_types_encoder::SolValue::abi_encode(&cctp_transer_cfg).into(),
+    )
+    .into_transaction_request()
+    .from(admin);
+
+    let cctp_rx = async_run!(rt, eth_client.execute_tx(cctp_tx).await.unwrap());
+
+    let cctp_address = cctp_rx.contract_address.unwrap();
+    info!("CCTP Transfer deployed at: {cctp_address}");
+
+    // approve the CCTP forwarder on deposit account
+    valence_e2e::utils::ethereum::valence_account::approve_library(
+        rt,
+        eth_client,
+        input_account,
+        cctp_address,
+    );
+
+    Ok(cctp_address)
 }
